pyproject.toml
[tool.poetry]
name = "neurodb"
version = "0.1.0"
description = "A database of shared resources"
authors = ["Neil Munro (NMunro)"]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
django-money = "^3.2.0"
pillow = "^10.0.0"
django-crispy-forms = "^2.0"
install = "^1.3.5"
crispy-bootstrap5 = "^0.7"
whitenoise = "^6.5.0"
python-decouple = "^3.8"
dj-database-url = "^2.0.0"
gunicorn = "^21.2.0"
django-phone-field = "^1.8.1"
psycopg = { version = ">=3.1", extras = ["binary"] }
cron-validator = "^1.0.8"
cron-converter = "^1.0.2"
poetry-plugin-export = "^1.6.0"


[tool.poetry.dev-dependencies]
coverage = "^7.1"
isort = "^5.10"
black = "^21.10b"
sphinx = "^4.2"
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
